towards #27: link reply processing marks TX slots now

+++ tests/BCLinkManagerTests.cpp
@@ -132,7 +132,7 @@ namespace TUHH_INTAIRNET_MCSOTDMA {
 				CPPUNIT_ASSERT_EQUAL(own_id, base_header->icao_id);
 				CPPUNIT_ASSERT_EQUAL(own_id, packet->getOrigin());
 				CPPUNIT_ASSERT_EQUAL(SYMBOLIC_LINK_ID_BROADCAST, packet->getDestination());
-				CPPUNIT_ASSERT_EQUAL(ushort(1), base_header->length_next);
+				CPPUNIT_ASSERT_EQUAL(uint(1), base_header->length_next);
 				CPPUNIT_ASSERT_EQUAL(uint(0), base_header->offset);
 				CPPUNIT_ASSERT_EQUAL(uint(0), base_header->timeout);
 //				coutd.setVerbose(false);
@@ -153,7 +153,7 @@ namespace TUHH_INTAIRNET_MCSOTDMA {
 				L2Packet* packet = phy_layer->outgoing_packets.at(0);
 				CPPUNIT_ASSERT_EQUAL(SYMBOLIC_LINK_ID_BROADCAST, packet->getDestination());
 				auto* base_header = (L2HeaderBase*) packet->getHeaders().at(0);
-				CPPUNIT_ASSERT_EQUAL(ushort(1), base_header->length_next);
+				CPPUNIT_ASSERT_EQUAL(uint(1), base_header->length_next);
 				// A non-zero offset means we must've scheduled a next broadcast.
 				CPPUNIT_ASSERT(base_header->offset > 0);
 				CPPUNIT_ASSERT_EQUAL(uint(0), base_header->timeout);
